Teach the parser to be more explicit about unknown record types
By the current convention all upper case words sarting a line denote a specific recordtype. This change will make the parser inform the user if an unknown record type is encountered, gracefully, in that the parser returns so that in a batch job other log files will still be processed. This commit also implements a stricter check for when/where event ide user variables are expected and will be parsed.